Create VADER Function
Adds a link to a google Jupiter notebook that creates a dataframe with VADER scores for News API Results for the three days before the target date we are interested in. The user needs to manually enter their News API Key, which is easier than hosting it somewhere on Google's server.